Name of Section component changed, Second section component, Button deleted
