--- src/Components/Second-Section/Second-Section.css
@@ -1,22 +1,0 @@
-.main-text {
-    font-size: 1.1rem;
-}
-
-.card-style {
-    width: 18rem;
-}
-
-.card {
-    width: 22.5rem;
-    height: auto;
-}
-
-.img {
-    height: 14rem;
-}
-
-.photo-animation:hover{
-    transform: scale(1.05);
-}
-
-
